change get surround location
